// qk-processing/src/credentials/type/fields.encrypt.type.ts
export type FieldsEncryptType = {
    certificateId: string
    graduatedName: string
    authenticatedBy: string
    qualificationName: string
    majors: string
    minors: string
    authenticatedTitle: string
    awardingInstitution: string
    qualificationLevel: string
    awardLevel: string
    studyLanguage: string
    info: string
    gpaFinalGrade: string
    authenticatedAt: Date
    studyStartedAt: Date
    studyEndedAt: Date
    graduatedAt: Date
    expiresAt: Date
}